+++ src/main/kotlin/com/terraformation/backend/tracking/event/Events.kt
@@ -10,6 +10,7 @@ import com.terraformation.backend.tracking.model.ExistingObservationModel
 import com.terraformation.backend.tracking.model.ExistingPlantingSiteModel
 import com.terraformation.backend.tracking.model.ReplacementDuration
 import com.terraformation.backend.tracking.model.ReplacementResult
+import java.math.BigDecimal
 import java.time.LocalDate
 
 /** Published when an organization requests that a monitoring plot be replaced in an observation. */
@@ -117,7 +118,7 @@ data class PlantingSiteMapEditedEvent(
 )
 
 data class T0SpeciesDensityAssignedEvent(
-    val plotDensity: Int,
+    val plotDensity: BigDecimal,
     val monitoringPlotId: MonitoringPlotId,
     val speciesId: SpeciesId,
 )
